+++ src/toc/toc-insert-windows.spec.ts
@@ -7,7 +7,7 @@ import { ITocService } from './toc.interface';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-describe('toc', () => {
+describe('toc - windows', () => {
   describe('insertToc', () => {
     let toc: Toc;
     let mdService: IMarkdown;
@@ -21,19 +21,19 @@ describe('toc', () => {
 
     let spyUpdateMd: jest.SpyInstance;
     const expectedContent =
-      '# Test\n\n' +
-      '<!-- toc -->\n' +
-      '- [Heading 1](#heading-1)\n' +
-      '  - [Subheading](#subheading)\n' +
-      '- [Heading 2](#heading-2)\n' +
-      '<!-- tocstop -->\n\n' +
-      '## Heading 1\n\n' +
-      '### Subheading\n\n' +
-      '## Heading 2\n';
+      '# Test\r\n\r\n' +
+      '<!-- toc -->\r\n' +
+      '- [Heading 1](#heading-1)\r\n' +
+      '  - [Subheading](#subheading)\r\n' +
+      '- [Heading 2](#heading-2)\r\n' +
+      '<!-- tocstop -->\r\n\r\n' +
+      '## Heading 1\r\n\r\n' +
+      '### Subheading\r\n\r\n' +
+      '## Heading 2\r\n';
     const expectedErrorMessage =
-      'Could not find placeholder\n' +
-      '<!-- toc -->\n' +
-      '<!-- tocstop -->\n' +
+      'Could not find placeholder\r\n' +
+      '<!-- toc -->\r\n' +
+      '<!-- tocstop -->\r\n' +
       'A toc update or insertion was not possible. Please sure the placeholder are set.';
 
     beforeEach(() => {
@@ -43,20 +43,20 @@ describe('toc', () => {
     });
 
     it('should add toc in placeholders', () => {
-      toc.filePath = 'fixtures/insert-with-placeholders.md';
+      toc.filePath = 'fixtures/insert-with-placeholders.windows.md';
       toc.insertToc();
-      expect(spyUpdateMd).toHaveBeenCalledWith('fixtures/insert-with-placeholders.md', expectedContent);
+      expect(spyUpdateMd).toHaveBeenCalledWith('fixtures/insert-with-placeholders.windows.md', expectedContent);
     });
 
     it('should update toc in placeholders', () => {
-      toc.filePath = 'fixtures/insert-with-outdated-toc.md';
+      toc.filePath = 'fixtures/insert-with-outdated-toc.windows.md';
       toc.insertToc();
-      expect(spyUpdateMd).toHaveBeenCalledWith('fixtures/insert-with-outdated-toc.md', expectedContent);
+      expect(spyUpdateMd).toHaveBeenCalledWith('fixtures/insert-with-outdated-toc.windows.md', expectedContent);
     });
 
     it('should throw error if placeholder are in one line', () => {
       let actualErrorMessage = '';
-      toc.filePath = 'fixtures/insert-with-placeholder-in-one-line.md';
+      toc.filePath = 'fixtures/insert-with-placeholder-in-one-line.windows.md';
       try {
         toc.insertToc();
       } catch (err) {
@@ -67,7 +67,7 @@ describe('toc', () => {
 
     it('should throw error if only toc placeholder exists', () => {
       let actualErrorMessage = '';
-      toc.filePath = 'fixtures/insert-without-stop-placeholder.md';
+      toc.filePath = 'fixtures/insert-without-stop-placeholder.windows.md';
       try {
         toc.insertToc();
       } catch (err) {
@@ -78,7 +78,7 @@ describe('toc', () => {
 
     it('should throw error if only tocstop placeholder exists', () => {
       let actualErrorMessage = '';
-      toc.filePath = 'fixtures/insert-without-start-placeholder.md';
+      toc.filePath = 'fixtures/insert-without-start-placeholder.windows.md';
       try {
         toc.insertToc();
       } catch (err) {