+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.3.0 (25.02.2021)
+
+- support windows carriage return and adding test cases
+
 ## 0.2.2 (24.02.2021)
 
 - add keywords, repository and homepage to package.json